+++ Electron/vue2/src/TUIRoom/components/Chat/index.vue
@@ -7,9 +7,7 @@
 
 <script setup lang="ts">
 import MessageList from './MessageList.vue';
-
 import ChatEditor from './ChatEditor/index.vue';
-
 import isMobile from '../../utils/useMediaValue';
 </script>