+++ Electron/vue2/src/TUIRoom/components/ManageMember/MemberItem/index.vue
@@ -11,9 +11,7 @@
 
 <script setup lang="ts">
 import MemberInfo from '../MemberItemCommon/MemberInfo.vue';
-
 import MemberControl from '../MemberControl/index.vue';
-
 import { UserInfo } from '../../../stores/room';
 import useMemberItem from './useMemberItemHooks';