+++ Electron/vue2/src/TUIRoom/hooks/useRoomEngine.ts
@@ -1,6 +1,8 @@
 import TUIRoomEngine from '@tencentcloud/tuiroom-engine-electron';
+import { isVue3 } from '../utils/constants';
 
 const roomEngine: Record<string, TUIRoomEngine | null> = { instance: null };
+const vueVersion = isVue3 ? 'vue3' : 'vue2';
 
 export default function useGetRoomEngine() {
   return roomEngine;
@@ -8,4 +10,11 @@ export default function useGetRoomEngine() {
 
 TUIRoomEngine.once('ready', () => {
   roomEngine.instance = new TUIRoomEngine();
+  roomEngine.instance?.callExperimentalAPI(JSON.stringify({
+    api: 'setFramework',
+    params: {
+      component: 'TUIRoomKit',
+      language: vueVersion
+    }
+  }))
 });