+++ Electron/vue3/packages/renderer/src/TUIRoom/components/ManageMember/index.vue
@@ -45,9 +45,7 @@
 
 <script setup lang='ts'>
 import { storeToRefs } from 'pinia';
-/// @TUIRoom-PlatformAdapter-Start
 import MemberItem from '../ManageMember/MemberItem/index.vue';
-/// @TUIRoom-PlatformAdapter-End
 import useGetRoomEngine from '../../hooks/useRoomEngine';
 import SvgIcon from '../common/SvgIcon.vue';
 import { useRoomStore } from '../../stores/room';