+++ Electron/vue3/packages/renderer/src/TUIRoom/components/RoomFooter/AudioControl.vue
@@ -64,9 +64,7 @@
 import { ref, onMounted, Ref, onUnmounted, computed } from 'vue';
 import { storeToRefs } from 'pinia';
 import { ElMessageBox, ElMessage } from '../../elementComp';
-/// @TUIRoom-PlatformAdapter-Start
 import Dialog from '../../elementComp/Dialog/index.vue';
-/// @TUIRoom-PlatformAdapter-End
 import IconButton from '../common/IconButton.vue';
 import AudioSettingTab from '../base/AudioSettingTab.vue';
 import { useRoomStore } from '../../stores/room';